refactor : refactor telegraf
