+++ java-module/mc-o11y-manager/src/main/java/com/mcmp/o11ymanager/facade/TelegrafConfigFacadeService.java
@@ -1,6 +1,6 @@
 package com.mcmp.o11ymanager.facade;
 
-import com.mcmp.o11ymanager.oldService.domain.interfaces.FileService;
+import com.mcmp.o11ymanager.service.interfaces.FileService;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
